+++ python/ferrmion/encode/ternary_tree.py
@@ -55,12 +55,12 @@ class TernaryTree(FermionQubitEncoding):
             n_modes (int): How many fermionic modes in the encoding.
             root_node (TTNode): The root node of the tree.
         """
+        self._enumeration_scheme = {}
         self.n_modes = n_modes
         self.n_qubits = n_modes
-        self.root = root_node
-        self.root.label = ""
+        self.root_node = root_node
+        self.root_node.root_path = ""
         self.vacuum_state = np.array([0] * self.n_qubits, dtype=np.uint8)
-        self._enumeration_scheme = {}
         super().__init__(self.n_modes, self.n_qubits)
 
     @classmethod
@@ -108,8 +108,8 @@ class TernaryTree(FermionQubitEncoding):
         """
         logger.debug("Setting enumeration scheme.")
         error_string = ""
-        if set(self.root.child_strings) != set(enumeration_dict.keys()):
-            error_string += f"Enumeration scheme {enumeration_dict} must contain all nodes {self.root.child_strings}.\n"
+        if set(self.root_node.child_strings) != set(enumeration_dict.keys()):
+            error_string += f"Enumeration scheme {enumeration_dict} must contain all nodes {self.root_node.child_strings}.\n"
 
         modes = set()
         qubits = set()
@@ -143,12 +143,21 @@ class TernaryTree(FermionQubitEncoding):
             {"": (0,0), "z": (1,1), "zz": (2,2)}
         """
         logger.debug("Setting default enumeration scheme")
-        logger.debug("Child strings %s", self.root.child_strings)
-        return {child: (i, i) for i, child in enumerate(self.root.child_strings)}
+        logger.debug("Child strings %s", self.root_node.child_strings)
+        enumeration_scheme = {}
+        child_labels = self.root_node.child_qubit_labels
+        spare_labels: set[int] = set(range(len(child_labels))).difference(
+            child_labels.values()
+        )
+        for child, label in child_labels.items():
+            if label is None:
+                label = spare_labels.pop()
+            enumeration_scheme[child] = (int(label), int(label))
+        return enumeration_scheme
 
     def as_dict(self):
         """Return the tree structure as a dictionary."""
-        return self.root.as_dict()
+        return self.root_node.as_dict()
 
     def add_node(self, node_string: str) -> "TernaryTree":
         """Add a node to the tree.
@@ -170,12 +179,12 @@ class TernaryTree(FermionQubitEncoding):
         if not valid_string:
             raise ValueError("Branch string can only contain x,y,z")
 
-        node = self.root
+        node = self.root_node
         for char in node_string:
             if isinstance(getattr(node, char), TTNode):
                 node = getattr(node, char)
             else:
-                node = node.add_child(char, node.label + char)
+                node = node.add_child(char, root_path=f"{node.root_path}{char}")
         return self
 
     @property
@@ -201,7 +210,7 @@ class TernaryTree(FermionQubitEncoding):
         """
         logger.debug("Building branch operator map for TernaryTree.")
 
-        branches = self.root.branch_strings
+        branches = self.root_node.branch_strings
 
         node_indices = {
             node: qubit for node, (_, qubit) in self.enumeration_scheme.items()
@@ -209,9 +218,9 @@ class TernaryTree(FermionQubitEncoding):
         branch_operator_map = {}
         for branch in branches:
             branch_operator_map[branch] = ["I"] * self.n_qubits
-            node = self.root
+            node = self.root_node
             for char in branch:
-                node_index = node_indices[node.label]
+                node_index = node_indices[node.root_path]
                 branch_operator_map[branch][node_index] = char.upper()
                 node = getattr(node, char, None)
             branch_operator_map[branch] = "".join(branch_operator_map[branch])
@@ -234,11 +243,11 @@ class TernaryTree(FermionQubitEncoding):
             {'': ('xzz', 'y'), 'x': ('xx', 'xy'), 'xz': ('xzx', 'xzy')}
         """
         logger.debug("Building string pairs for TernaryTree.")
-        node_set = self.root.child_strings
+        node_set = self.root_node.child_strings
 
         pairs = {}
         for node_string in node_set:
-            node = self.root
+            node = self.root_node
             for char in node_string:
                 node = getattr(node, char)
 
@@ -257,9 +266,9 @@ class TernaryTree(FermionQubitEncoding):
                         break
 
             if x_string.count("y") % 2 == 0:
-                pairs[node.label] = x_string, y_string
+                pairs[node.root_path] = x_string, y_string
             elif y_string.count("y") % 2 == 0:
-                pairs[node.label] = y_string, x_string
+                pairs[node.root_path] = y_string, x_string
 
         return pairs